Started case 1, exploration 1
I decided to start working with matterjs and p5, using the superiority theory as my motivating material. I had the idea of generating an image pixel by pixel, and then be able to destroy it. I thought this could maybe be funny in a mean way? Feeling superior to things or people. Thinking about that, I had the idea of using a fakeFace API to generate random people that don't exist as destrution photos. I didn't manage to make that work for now but i have the pixel thing going. Will try tomorrow to make the API work. I think the pixels appearing one after the other already has a something that is inherently funny. Maybe it's because the system seems slow and weak and old, which kind of related to the theory. The face appearing pixelated also has something a bit funny about it. Visually it makes it funny but it makes it also maybe ok to laugh about it because it is sort of anonymous? I like it, my goal is to maybe finish this small sketch by making it explode in some way... maybe using a button or something. Yeah dunno if I made that clear but each pixel is also a matterjs physical body so it can be destroyed, sort of like angry birds ish. So. Would be nice to have a bunch of these experiments in order to better understand how to make a bigger case.