Initiated image segmentation with demo from MediaPipe website
So I experimented I just uploaded the example demo from the MediaPipe webpage image segmentation, where you can click on an image and it segments the thing that it thinks is important I guess. I switched for bird images (was cat and dog before) and it worked well but cut off the legs of one of the birds. I did some research and found out that I could switch the model for one that is trained better on birds and that could be better at segmenting birds. I ended up on a tensorflow webpage (https://www.npmjs.com/package/@tensorflow-models/tasks) where they have an example using different models with runtimes. I tried all three into my mediaPipe example and one of them worked in getting all the bird on the image including the legs so I'll keep this one for now.
I also tried the webcam option at the bottom, and switched the webcam to a video input using video footage of birds in slow motion from youtube, and its pretty laggy. Since there is a latency in the segmentation it's always a bit late in detecting/segmeting so I don't know if I want to dive into this for now. I've been looking for other models specifically trained on birds. Will look more into this.
I think my computer has a hard time processing all of this which makes me frustrated because comedy is all about TIMING, and I feel like if I want to program comedy my developing environment has to be sharp. Maybe there's something interesting in talking about latency and performance in relation to humour and timing. Also maybe I could use this to make humour. I had a conversation with a collegue about a previous project I was working on that I presented in a sort-of exhibition context, and it wasnt working perfectly and that was annoying to me because I knew what I wanted to offer as an experience. But she told me that the bugs occuring were the reason why it was funny and enaging. It made me think a lot. I think surprise is part of why people laugh, and its also a never ending source of laughter, in contrast to like a predetermined programmed joke. Once you experience it ONCE then maybe you're done with it. I wonder if I can make something where bugs are part of it and are not annoying to me as a designer with a VISION.
So yeah, at this point of commit, I did not do much except having a "working" example of bird segmentation that I will probably use for this project. My plan is now to get a working example of object detection using video footage of birds. I think the segmentation part will be more accurate if I detect a bird first, cut out the rest of the image and segment from this cut out image.