Bird detection working!
So I have the bird detection working with cute little rectangles around the detected birds. AMAZING.
It really gives me a feel of like detective surveillance, like those 2000' movies with glasses that could detect people and I think it would be interesting to go in that direction. Like scan for birds and then check if there is a match with murderer? Or give like random characteristics to the bird after anaylising against the database of criminals. I LIKE IT.