Implemented image segmentation
At this point I have a video in an html element of birds looping on itself, when clicking and holding mouse down, MediaPipe runs object detection on video feed. Red boxes appear around the detections (birds in the sky). When releasing mouse button, the detection stops, but when holding mouse down and pressing "s" key, a new image is created from the bounding box around the first bird currently being detected. Then this "cut-out" image of the bird is used for the segmentation from its background. Then a new image is created, and using the segemtation mask, the pixels inside the mask are copied from the cut out image and the pixels as background (outside the mask) are transparent. The result is a copy of the bird on a transparent background.

So the first part of this is now done. I now need to make a good working loop with a base of an interface. I also want to see how I can make generative characteristics based on the data from the segmentation? I need to think more about it. But yeah happy about this specific part being done. Feels like the rest will be understanding how to put everything together with visuals and sounds to make it make more sense and figure out next steps.