Finished sketch idea of copying birds detected in video feed and drawing them on separate canvas to make a trail
for this sketch, I have a two canvases, one displaying a video feed of flying birds and the second one is a copy of the detections leaving a trail of their movement. On the top one, the birds have a lable with the percentage of how accurate the model is that a bird is being detected. I decided to put this because I felt there was a disconnect between the two canvases. I think now there is a clearer correlation between both, and that there is a detection process being applied to the top one. Before I had put a rectangle outline around the detection of the birds, and it looked interesting but it was a chaotic visually when many birds were detected. Also, I like the contrast of something more fluid and organic at the top and something more digital at the bottom. For now i'm happy with this result, I think it could be interesting to test with a live camera feed somewhere.