Added a trail for the copied birds! Looks real nice.
I feel like this is not really inline with making something FUNNY per say, but i guess it is a sort of result of using humour theory to design interactive experiments. I mean will all experiments become funny? No, the goal is not explore how to make something funny, because i'm not really interested in what is funny and what is not. But I need to ponder a bit more on that.