New branch new experiment!
So I deciced to try another experiment more around my own thesis where birds are ridiculous so they become funny??? so right now i have my video feed of birds and then at the bottom, i have a circle showing random birds in a fixed dimension so it distorts some of them. I have this idea because I did an error while programming and it was displaying the birds in a sort of weird and chaotic way where the size of the image was not adapted to the original size of the capture. I thought again about that humour comes from surprise and bugs and i HATE THAT because that means i have no control and i have to mess things up and its making me anxious. Its like comedy can arise or not depending on what happens. So yeah again = people think the bugs are funny which should be freeing for me but it is not.
ARG. arrrrrg. I think i'm almost close to going with this. yes. i need to let go.