Added tunnel enclosure and collision filters
I created two big static rectangle on each side of the sketch with a space in between, creating a sort of tunnel through wich the anal beads are stuck inside. I created this because the player could click on a bead and drag it to the side, making the experience less realistic. The toy would then just come out entirely from the side. I thought, in the real world, you have to pull straightforward to get the ball out, not from the sides. And also, in the real world it is a real person and not two ellipses as cheeks, so the toy would not come out otherwise. I felt like I needed to constrain the ball tightly to force the player to pull downards only. But then the toy became really static and the chain like effect was not interesting. I decided to make a treshold where past a certain point on the y axis, 3/4 of the canvas about, the collision did not apply anymore between the tunnel enclosures and the beads, meaning that they would be tightly constrained when coming out initially but when pulled to the point of letting the next one out, the collision would be deactivated. So now I have the butter and the money of the butter. YAS.