Switched back to Paul Dinning video, detections on pause, birdcam on the video
So as I was working with many videos of birds it started to feel weird and too complicated. I think for my toys to work and feel good to me they need to be quite simple. And having many videos of birds that don't all work properly was weird. I decided to go back to my original beautiful video of Paul Dinning. I like that it is sort of a moving painting your are looking at, and less a tv wth changing channels. The thing is I need to associate a song with a bird now, because clicking on a bird doesnt switch the video anymore, and if you go back to a bird then if it doesn't have the same song, i guess the surprise is funny, and also the fact that it changed emotion fast. But i think it makes the program more funny to have consistency, that way you sort of feel like youre getting to know the birds personally? Like you can trust them. I like that while you are pausing, meaning you are on the video, the songs stay the same, but when you let the video go, then the song is different, which is sort of the best of both worlds, because in that frozen moment in time you don't really change emotion, but when the video starts moving again, then so are the thoughts and emotions. I guess that is more journal material tho.
I chose to freeze the video because it made it more simple for people to click on the birds, and also for performance issues. Because of that I can make the video bigger which is also nice. because it is bigger I also made the birdcam, the circles of birds, on top of the actual video instead of outside. I made it so that if the birds are on the left the cam is on the right and vice versa. it adds some dynamism to it. Over all i think all these choices are really helping the design both conceptually and formally.