Cursor change to magnifying glass, video pauses when hovering,
I decided to add a cursor magnifying glass when hovering over a bird and have the end of the zoom cone follow the mouse when inside the detection. I'm not sure if I like it like that yet. There's something about the amount of details that I feel have an impact on how funny something is. Even something like the cone following the mouse, adds a detail that the user notices and adds almost like a barrier or like a rule or I don't know, something for the brain to assimilate that makes the humour a bit less efficient? Maybe I worked for too long on it now and don't know anything anymore. I don't feel much about this so it makes it difficult to understand if the humour lives on.

The video pauses when the mouse is moving and continues when mouse stops moving. I chose this because I feel like it needs less explanation. The user will move and stop moving the mouse and will understand that it pauses or not the video. I'm looking for a way to not add any instructions. Again, hard to tell because I'm not testing it, but it feels right. Something about the pausing video though, I'm thinking it could maybe be interesting to add a bit of movement to the static image of the bird, like a smooth zoom in as the song is playing? I feel like its veryyyyy static now. i like the freeztime element and actually needed it to make the program not die, but now its maybe a bit too static?